Add files via upload
The code for this Repository is borrowed from the "https://github.com/tkipf/pygcn.git" repository. Code here is amazing and if you like the code support the writer of this code Thomas Kipf whose GitHub repository is given above. I just added my understanding of code for a beginner who might have a problem understanding the code might find it easier.